Initial typeparsing class model and suggested sql structure. I was looking at abc at first, but it seems like typeannotations are more popular from the pydantic documentation (and equivalent anyways...).
The basic idea is simple. When an "object in the universe" is referenced, the query will first go to base_model table (note: You can use a foreignkey directly to the other table, but you lose some flexibility in typing).

The base_model table would then hold a reference to the information for that type. So, if object A has a reference to object B, you could look up the object in the base_model table and do the necessary typing inferences.